+++ src/modules/core/ConfigVar.h
@@ -153,6 +153,7 @@ constexpr const char *VoxformatBinvoxVersion = "voxformat_binvoxversion";
 constexpr const char *VoxformatSkinApplyTransform = "voxformat_skinapplytransform";
 constexpr const char *VoxformatSkinAddGroups = "voxformat_skinaddgroups";
 constexpr const char *VoxformatSkinMergeFaces = "voxformat_skinmergefaces";
+constexpr const char *VoxformatMCSeparateWater = "voxformat_mcseparatewater";
 constexpr const char *VoxformatGMLRegion = "voxformat_gmlregion";
 constexpr const char *VoxformatGMLFilenameFilter = "voxformat_gmlfilenamefilter";
 constexpr const char *VoxformatOSMURL = "voxformat_osmurl";